backend_cloudinary/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
pymongo==4.6.1
bcrypt==4.1.2
python-dotenv==1.0.0
werkzeug==3.0.1
requests==2.31.0
gunicorn==21.2.0
cloudinary==1.38.0
